services/gen_testcase/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
google-genai
orjson
python-jose==3.4.0
ollama
